+++ codex-rs/config/BUILD.bazel
@@ -1,0 +1,6 @@
+load("//:defs.bzl", "codex_rust_crate")
+
+codex_rust_crate(
+    name = "config",
+    crate_name = "codex_config",
+)
